WIP: analyze tool swapped
