+++ tools/__init__.py
@@ -3,7 +3,6 @@ Tool implementations for Zen MCP Server
 """
 
 from .analyze import AnalyzeTool
-from .analyzeworkflow import AnalyzeWorkflowTool
 from .chat import ChatTool
 from .codereview import CodeReviewTool
 from .consensus import ConsensusTool
@@ -21,7 +20,6 @@ __all__ = [
     "CodeReviewTool",
     "DebugIssueTool",
     "AnalyzeTool",
-    "AnalyzeWorkflowTool",
     "ChatTool",
     "ConsensusTool",
     "ListModelsTool",
