+++ tools/workflow/workflow_mixin.py
@@ -932,7 +932,7 @@ class BaseWorkflowMixin(ABC):
         """
         Get the message to show when work is complete.
         Tools can override for custom messaging.
-        
+
         Args:
             expert_analysis_used: True if expert analysis was successfully executed
         """
@@ -943,26 +943,26 @@ class BaseWorkflowMixin(ABC):
             "reasoning that led to this conclusion—make it easy for a developer to understand exactly where "
             "the problem lies."
         )
-        
+
         # Add expert analysis guidance only when expert analysis was actually used
         if expert_analysis_used:
             expert_guidance = self.get_expert_analysis_guidance()
             if expert_guidance:
                 return f"{base_message}\n\n{expert_guidance}"
-        
+
         return base_message
 
     def get_expert_analysis_guidance(self) -> str:
         """
         Get additional guidance for handling expert analysis results.
-        
+
         Subclasses can override this to provide specific instructions about how
         to validate and use expert analysis findings. Returns empty string by default.
-        
+
         When expert analysis is called, this guidance will be:
         1. Appended to the completion next steps message
         2. Added as "important_considerations" field in the response data
-        
+
         Example implementation:
         ```python
         def get_expert_analysis_guidance(self) -> str:
@@ -973,7 +973,7 @@ class BaseWorkflowMixin(ABC):
                 "recommendations align with the codebase context."
             )
         ```
-        
+
         Returns:
             Additional guidance text or empty string if no guidance needed
         """
@@ -1070,7 +1070,7 @@ class BaseWorkflowMixin(ABC):
             else:
                 # Expert analysis was successfully executed - include expert guidance
                 response_data["next_steps"] = self.get_completion_next_steps_message(expert_analysis_used=True)
-                
+
                 # Add expert analysis guidance as important considerations
                 expert_guidance = self.get_expert_analysis_guidance()
                 if expert_guidance:
